Rename research_final.py to Airbnb price prediction research.py
Airbnb provides some general guidance to the hosts for pricing the listings however, there is no easy way to determine the optimum price. Although there are some websites available for this purpose like beyondpricing.com that comes at a high cost. 

This project aims to identify the key features of a listing by the use of machine learning models that would assist the hosts in determining the optimum price of their listings. We have analyzed over 75K listings of California to identify which features, amenities, and experiences offered (based on the description) are most useful in predicting the optimum listing price.